+++ terratex_reallife/ENVIRO/ruhezone/ruhezone_server.lua
@@ -1,10 +1,11 @@
 ruhezonen = {}
 
 function createAndLoadRuheZonen()
+    ruhezonen = {};
     local result = MySql.helper.getSync("objects_restareas", "*");
     for theKey, dasatz in ipairs(result) do
-        local area = createRadarArea(tonumber(dasatz["leftX"]), tonumber(dasatz["buttonY"]), tonumber(dasatz["sizeX"]), tonumber(dasatz["sizeY"]), 0, 200, 0, 150, getRootElement())
-        ruhezonen[tonumber(dasatz["ID"])] = area
+        local area = createRadarArea(dasatz.leftX, dasatz.buttonY, dasatz.sizeX, dasatz.sizeY, 0, 200, 0, 150, getRootElement())
+        ruhezonen[dasatz.ID] = area
     end
 end
 addEventHandler("onResourceStart", getResourceRootElement(getThisResource()), createAndLoadRuheZonen)
@@ -23,16 +24,18 @@ end
 
 function setRZoneCorner_func(thePlayer, cmd, corner)
     if not corner then
-        outputChatBox("Spezify Corner with 1 or 2 (/setrzonecorner 1|2)");
+        showError(thePlayer, "Spezify Corner with 1 or 2 (/setrzonecorner 1|2)");
     end
-    if corner ~= 1 and corner ~= 2 then
-        outputChatBox("Spezify Corner with 1 or 2 (/setrzonecorner 1|2)");
+    if corner ~= "1" and corner ~= "2" then
+        showError(thePlayer, "Spezify Corner with 1 or 2 (/setrzonecorner 1|2)");
     end
 
     if (isAdminLevel(thePlayer, 3)) then
         local posX, posY, posZ = getElementPosition(thePlayer)
         local posXString = "posX" .. corner;
         local posYString = "posY" .. corner;
+
+        outputChatBox("Position saved!", thePlayer);
 
         vioSetElementData(thePlayer, posXString, posX);
         vioSetElementData(thePlayer, posYString, posY);
@@ -51,15 +54,15 @@ function addRuheZone(thePlayer)
                 posXA = vioGetElementData(thePlayer, "posX1")
                 posXB = vioGetElementData(thePlayer, "posX2")
             else
-                posXA = vioGetElementData(thePlayer, "posX1")
-                posXB = vioGetElementData(thePlayer, "posX2")
+                posXA = vioGetElementData(thePlayer, "posX2")
+                posXB = vioGetElementData(thePlayer, "posX1")
             end
             if (vioGetElementData(thePlayer, "posY1") < vioGetElementData(thePlayer, "posY2")) then
                 posYA = vioGetElementData(thePlayer, "posY1")
                 posYB = vioGetElementData(thePlayer, "posY2")
             else
-                posYA = vioGetElementData(thePlayer, "posY1")
-                posYB = vioGetElementData(thePlayer, "posY2")
+                posYA = vioGetElementData(thePlayer, "posY2")
+                posYB = vioGetElementData(thePlayer, "posY1")
             end
 
             local sizeX = posXB - posXA
@@ -74,7 +77,8 @@ function addRuheZone(thePlayer)
 
             outputChatBox("Ruhezone ID " .. ID .. " created", thePlayer, 255, 0, 0)
             local area = createRadarArea(posXA, posYA, sizeX, sizeY, 0, 200, 0, 150, getRootElement())
-            ruhezonen[ID] = area
+            ruhezonen[tonumber(ID)] = area
+
             triggerClientEvent(getRootElement(), "empfangeRuhezonenData", thePlayer, ruhezonen)
 
         else
@@ -87,9 +91,10 @@ addCommandHandler("addruhe", addRuheZone, false, false)
 function destroyRuheZone(thePlayer, cmd, ID)
     if (isAdminLevel(thePlayer, 3)) then
         if tonumber(ID) then
+            ID = tonumber(ID);
             if (ruhezonen[ID]) then
-                destroyElement(ruhezonen[ID])
-                ruhezonen[ID] = false
+                destroyElement(ruhezonen[ID]);
+                ruhezonen = table.removeKey(ruhezonen, ID)
                 triggerClientEvent(getRootElement(), "empfangeRuhezonenData", thePlayer, ruhezonen)
                 MySql.helper.delete("objects_restareas", { ID = ID });
                 outputChatBox("Ruhezone gelöscht", thePlayer, 255, 0, 0)