+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 * [ADDED] Sql Update 1.0.0 File
 * [ADDED] #16 Table Creation for Teleportmarkers
 * [ADDED] #16 New Teleportmarkers Script
+* [FIXED] #17 Fixed Ruhezonen Bugs
+* [FIXED] #20 /hkasse was working everywhere
 * [FIXED] #19 Fixed Afk Bug of the job "Meeresreiniger"
 * [FIXED] #18 Hotdogsellers can now set a prices that is smaller or equal then the money of the receiver
 * [CHANGED] #17 Rename Commands of restareas to /setRZoneCorner and /delRZone